Tests/LibWeb/Text/expected/HTML/parser-streams-utf8-split.txt

contains replacement: false
length: 4
